# wydgo748/api/openapi.yaml
openapi: 3.1.0
info:
  title: WYD administrative API (local-first)
  version: 0.2.0
paths:
  /api/v1/staff/players/{uid}/teleport:
    post:
      summary: Teleport the selected live session to a public city (moderation.player.teleport); embedded World only
      security: [{ staffSession: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRF'
        - name: uid
          in: path
          required: true
          schema: { type: string }
          description: Must equal the UID in the command body
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeleportCommand' }
      responses:
        '200':
          description: Domain receipt; retry uncertain outcomes with the identical body. No arbitrary coordinates; rejected in instances, war arenas, trade or ghost shop. Persistence failure rolls back before publication.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeleportResult' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/staff/players/{uid}/kick:
    post:
      summary: Disconnect the selected live session (moderation.player.kick); embedded World only
      security: [{ staffSession: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRF'
        - name: uid
          in: path
          required: true
          schema: { type: string }
          description: Must equal the UID in the command body
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KickCommand' }
      responses:
        '200':
          description: Domain receipt; retry uncertain outcomes with the identical body. Does not ban. Persistence failure is reported separately.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KickResult' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/staff/events/global-drop:
    get:
      summary: Inspect process-local global drop event (game.event.global-drop capability)
      security: [{ staffSession: [] }]
      responses:
        '200':
          description: Authoritative World snapshot; event is not resumed after restart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GlobalDropStatus' }
        default: { $ref: '#/components/responses/Error' }
    post:
      summary: Start or stop global drop through embedded World queue; staff session rechecked at execution
      security: [{ staffSession: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GlobalDropCommand' }
      responses:
        '200':
          description: Domain receipt (code ok means success); retry uncertain outcomes with the identical body and operationId
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GlobalDropResult' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/staff/events/quiz:
    get:
      summary: Inspect process-local quiz event (game.event.quiz capability)
      security: [{ staffSession: [] }]
      responses:
        '200':
          description: Authoritative World snapshot; event is not resumed after restart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizStatus' }
        default: { $ref: '#/components/responses/Error' }
    post:
      summary: Start or stop quiz through embedded World queue; staff session rechecked at execution
      security: [{ staffSession: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuizCommand' }
      responses:
        '200':
          description: Domain receipt (code ok means success); retry uncertain outcomes with the identical body and operationId
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizResult' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/staff/bosses:
    get:
      summary: Inspect configured boss lifecycle (game.boss.summon capability)
      security: [{ staffSession: [] }]
      responses:
        '200':
          description: Authoritative World snapshot of configured bosses
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BossesStatus' }
        default: { $ref: '#/components/responses/Error' }
    post:
      summary: Summon a configured dead boss through the embedded World queue; staff session rechecked at execution
      security: [{ staffSession: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BossesCommand' }
      responses:
        '200':
          description: Domain receipt; retry uncertain outcomes with the identical body and operationId
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BossesResult' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/session:
    get:
      summary: Get session and CSRF token; create an anonymous session if absent
      responses:
        '200':
          description: Current session; never a bearer credential
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/login:
    post:
      summary: Rotate session after game password, staff allowlist and configured administrative PIN verification
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [username, password, adminPin]
              properties:
                username: { type: string, minLength: 1, maxLength: 12 }
                password: { type: string, minLength: 1, maxLength: 12, writeOnly: true }
                adminPin: { type: string, pattern: '^[0-9]{6,12}$', writeOnly: true }
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/logout:
    post:
      parameters:
        - $ref: '#/components/parameters/CSRF'
      responses:
        '200':
          description: Session invalidated, cookie expired
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, const: signed_out }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/staff/overview:
    get:
      summary: One fresh World snapshot; requires both read capabilities
      security: [{ staffSession: [] }]
      parameters:
        - in: query
          name: search
          schema: { type: string, maxLength: 48 }
          description: Case-insensitive name, account or UID substring; maximum 48 Unicode characters
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, maximum: 10000, default: 0 }
      responses:
        '200':
          description: Snapshot sorted by UID, maximum 50 players; pages are not a stable cursor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Overview' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/staff/players/{uid}/persistent:
    get:
      summary: Read one persisted character projection from PostgreSQL
      security: [{ staffSession: [] }]
      parameters:
        - in: path
          name: uid
          required: true
          schema: { type: string, minLength: 1, maxLength: 64 }
          description: Canonical persistent character UID from the live snapshot
        - in: query
          name: account
          required: true
          schema: { type: string, minLength: 1, maxLength: 12 }
          description: Account owning the character in the live snapshot
      responses:
        '200':
          description: Independent PostgreSQL read; asOf is not the World snapshot timestamp
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PersistentPlayer' }
        default: { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    staffSession:
      type: apiKey
      in: cookie
      name: __Host-wyd_session
      description: HttpOnly Secure SameSite=Lax. Local dev uses wyd_local_session without Secure.
  parameters:
    CSRF:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
      description: Token from session endpoint. POST also requires exact Origin and session cookie.
  responses:
    Error:
      description: Rejected request (400/401/403/415), rate limit (429 with Retry-After), or unavailable dependency (503)
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
    TeleportCommand:
      type: object
      additionalProperties: false
      required: [version, operationId, epoch, uid, account, expectedSessionId, reason, destination]
      description: Bind UID, canonical account and sessionId from an Overview to its moderationEpoch. Receipts are actor/body-bound and retained up to 1024 per process without eviction. Session authorization is rechecked by World before execution.
      properties:
        version: { type: integer, const: 1 }
        operationId: { type: string, minLength: 16, maxLength: 64, pattern: '^[a-zA-Z0-9-]+$' }
        epoch: { type: string, pattern: '^[a-f0-9]{32}$' }
        uid: { type: string, minLength: 1, maxLength: 128 }
        account: { type: string, pattern: '^[a-z0-9]{1,12}$' }
        expectedSessionId: { type: integer, minimum: 1, maximum: 9007199254740991 }
        reason: { type: string, minLength: 3, maxLength: 200, description: Nonblank reason without control characters }
        destination: { type: string, enum: [armia, azran, erion, nippleheim] }
    TeleportResult:
      type: object
      required: [operationId, code, replayed, x, y]
      properties:
        operationId: { type: string }
        code: { type: string, enum: [ok, persistence_failed, player_busy, restricted_area, destination_unavailable, stale_target, invalid_command, capability_required, server_maintenance, operation_conflict, operation_capacity] }
        replayed: { type: boolean }
        x: { type: integer, minimum: 0, maximum: 4095, description: Final coordinate only when code is ok; otherwise zero }
        y: { type: integer, minimum: 0, maximum: 4095, description: Final coordinate only when code is ok; otherwise zero }
    KickCommand:
      type: object
      additionalProperties: false
      required: [version, operationId, epoch, uid, account, expectedSessionId, reason]
      description: Bind UID, canonical account and sessionId from an Overview to its moderationEpoch. Receipts are actor/body-bound and retained up to 1024 per process without eviction. Session authorization is rechecked by World before execution.
      properties:
        version: { type: integer, const: 1 }
        operationId: { type: string, minLength: 16, maxLength: 64, pattern: '^[a-zA-Z0-9-]+$' }
        epoch: { type: string, pattern: '^[a-f0-9]{32}$' }
        uid: { type: string, minLength: 1, maxLength: 128 }
        account: { type: string, pattern: '^[a-z0-9]{1,12}$' }
        expectedSessionId: { type: integer, minimum: 1, maximum: 9007199254740991 }
        reason: { type: string, minLength: 3, maxLength: 200, description: Nonblank reason without control characters }
    KickResult:
      type: object
      required: [operationId, code, replayed]
      properties:
        operationId: { type: string }
        code: { type: string, enum: [ok, disconnected_persistence_failed, stale_target, invalid_command, capability_required, server_maintenance, operation_conflict, operation_capacity] }
        replayed: { type: boolean }
    GlobalDropCommand:
      type: object
      additionalProperties: false
      required: [version, operationId, epoch, expectedEventId, action, reason, itemId, ratePercent, maxDrops, durationSeconds]
      description: Start requires itemId > 0, ratePercent 1..100 and at least one positive limit. Both limits stop at the first reached. Stop requires an event ID and zero numeric settings. Operation IDs are bound to actor and exact payload, retained up to 1024 commands per process plus one final stop; epoch prevents replay across restarts.
      properties:
        version: { type: integer, const: 1 }
        operationId: { type: string, minLength: 16, maxLength: 64, pattern: '^[a-zA-Z0-9-]+$' }
        epoch: { type: string, minLength: 32, maxLength: 32 }
        expectedEventId: { type: string, maxLength: 64 }
        action: { type: string, enum: [start, stop] }
        reason: { type: string, minLength: 3, maxLength: 200 }
        itemId: { type: integer, minimum: 0, maximum: 65535 }
        ratePercent: { type: integer, minimum: 0, maximum: 100 }
        maxDrops: { type: integer, minimum: 0, maximum: 10000000 }
        durationSeconds: { type: integer, minimum: 0, maximum: 604800 }
    GlobalDropStatus:
      type: object
      required: [version, epoch, asOf, eventId, active, itemId, itemName, ratePercent, maxDrops, dropped, startedAt, endsAt, endedBy, actor]
      properties:
        version: { type: integer, const: 1 }
        epoch: { type: string }
        asOf: { type: string, format: date-time }
        eventId: { type: string }
        active: { type: boolean }
        itemId: { type: integer }
        itemName: { type: string }
        ratePercent: { type: integer }
        maxDrops: { type: integer }
        dropped: { type: integer }
        startedAt: { type: [string, 'null'], format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        endedBy: { type: string, enum: ['', time, quantity, manual] }
        actor: { type: string }
    GlobalDropResult:
      type: object
      required: [operationId, code, replayed, status]
      properties:
        operationId: { type: string }
        code: { type: string, enum: [ok, invalid_command, capability_required, server_maintenance, stale_event, operation_conflict, operation_capacity, event_active, event_inactive, unknown_item, event_id_unavailable] }
        replayed: { type: boolean }
        status: { $ref: '#/components/schemas/GlobalDropStatus' }
    QuizCommand:
      type: object
      additionalProperties: false
      required: [version, operationId, epoch, expectedEventId, action, reason, itemId, rewardQuantity, durationSeconds]
      description: Start requires itemId > 0, quantity 1..60 and duration 60..86400 in whole minutes. First round is immediate, then every minute; answers expire after 10 seconds. Stop requires an event ID and zero settings. Retry uncertain outcomes with the identical body; bounded process-local receipts, epoch and event ID prevent replay and stale writes.
      properties:
        version: { type: integer, const: 1 }
        operationId: { type: string, minLength: 16, maxLength: 64, pattern: '^[a-zA-Z0-9-]+$' }
        epoch: { type: string, minLength: 32, maxLength: 32 }
        expectedEventId: { type: string, maxLength: 64 }
        action: { type: string, enum: [start, stop] }
        reason: { type: string, minLength: 3, maxLength: 200 }
        itemId: { type: integer, minimum: 0, maximum: 65535 }
        rewardQuantity: { type: integer, minimum: 0, maximum: 60 }
        durationSeconds: { type: integer, minimum: 0, maximum: 86400, multipleOf: 60 }
    QuizStatus:
      type: object
      required: [version, epoch, asOf, eventId, active, itemId, itemName, rewardQuantity, rounds, correct, rewarded, deliveryFailures, startedAt, endsAt, nextRoundAt, roundClosesAt, endedBy, actor]
      properties:
        version: { type: integer, const: 1 }
        epoch: { type: string }
        asOf: { type: string, format: date-time }
        eventId: { type: string }
        active: { type: boolean }
        itemId: { type: integer }
        itemName: { type: string }
        rewardQuantity: { type: integer }
        rounds: { type: integer }
        correct: { type: integer }
        rewarded: { type: integer }
        deliveryFailures: { type: integer }
        startedAt: { type: [string, 'null'], format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        nextRoundAt: { type: [string, 'null'], format: date-time }
        roundClosesAt: { type: [string, 'null'], format: date-time }
        endedBy: { type: string, enum: ['', time, manual] }
        actor: { type: string }
    QuizResult:
      type: object
      required: [operationId, code, replayed, status]
      properties:
        operationId: { type: string }
        code: { type: string, enum: [ok, invalid_command, capability_required, server_maintenance, stale_event, operation_conflict, operation_capacity, event_active, event_inactive, unknown_item, event_id_unavailable] }
        replayed: { type: boolean }
        status: { $ref: '#/components/schemas/QuizStatus' }
...
    BossesStatus:
      type: object
      required: [version, epoch, asOf, bosses]
      properties:
        version: { type: integer, const: 1 }
        epoch: { type: string, minLength: 32, maxLength: 32 }
        asOf: { type: string, format: date-time }
        bosses:
          type: array
          items: { $ref: '#/components/schemas/BossStatus' }
    BossesResult:
      type: object
      required: [operationId, code, replayed, status]
      properties:
        operationId: { type: string }
        code: { type: string, enum: [ok, invalid_command, capability_required, server_maintenance, stale_boss, operation_conflict, operation_capacity, unknown_boss, boss_alive, spawn_failed] }
        replayed: { type: boolean }
        status: { $ref: '#/components/schemas/BossesStatus' }
...